chore(data-models): Update TreeView to fix TS errors
